+++ src/Libraries/SmartStore.Data/Migrations/201705281903241_MoreIndexes.cs
@@ -1,13 +1,19 @@
 namespace SmartStore.Data.Migrations
 {
-	using System;
 	using System.Data.Entity.Migrations;
-	using Setup;
+	using System.Web.Hosting;
+	using Core.Data;
 
-	public partial class MoreIndexes : DbMigration, ILocaleResourcesProvider, IDataSeeder<SmartObjectContext>
+	public partial class MoreIndexes : DbMigration
 	{
 		public override void Up()
 		{
+			if (HostingEnvironment.IsHosted && DataSettings.Current.IsSqlServer)
+			{
+				// Avoid "Column 'Name' in table 'dbo.ProductVariantAttributeValue' is of a type that is invalid for use as a key column in an index".
+				Sql("If -1 = (SELECT CHARACTER_MAXIMUM_LENGTH FROM INFORMATION_SCHEMA.COLUMNS WHERE TABLE_NAME = 'ProductVariantAttributeValue' AND COLUMN_NAME = 'Name') ALTER TABLE [dbo].[ProductVariantAttributeValue] ALTER COLUMN [Name] nvarchar(4000) NULL;");
+			}
+
 			CreateIndex("dbo.Product_Category_Mapping", "IsFeaturedProduct");
 			CreateIndex("dbo.Product_Manufacturer_Mapping", "IsFeaturedProduct");
 			CreateIndex("dbo.SpecificationAttribute", "AllowFiltering");
@@ -26,23 +32,6 @@ namespace SmartStore.Data.Migrations
 			DropIndex("dbo.SpecificationAttribute", new[] { "AllowFiltering" });
 			DropIndex("dbo.Product_Manufacturer_Mapping", new[] { "IsFeaturedProduct" });
 			DropIndex("dbo.Product_Category_Mapping", new[] { "IsFeaturedProduct" });
-		}
-
-		public bool RollbackOnFailure
-		{
-			get { return false; }
-		}
-
-		public void Seed(SmartObjectContext context)
-		{
-			context.MigrateLocaleResources(MigrateLocaleResources);
-
-			context.SaveChanges();
-		}
-
-		public void MigrateLocaleResources(LocaleResourcesBuilder builder)
-		{
-			builder.AddOrUpdate("Common.For", "For: {0}", "Für: {0}");
 		}
 	}
 }